+++ src/extension/runner/index.ts
@@ -1,4 +1,3 @@
-import { GrpcTransport } from '@protobuf-ts/grpc-transport'
 import { DuplexStreamingCall } from '@protobuf-ts/runtime-rpc/build/types/duplex-streaming-call'
 import {
   type Pseudoterminal,
@@ -7,33 +6,25 @@ import {
   type TerminalDimensions,
   EventEmitter,
 } from 'vscode'
-import { RpcOptions, UnaryCall } from '@protobuf-ts/runtime-rpc'
 
 import type { DisposableAsync } from '../../types'
 import {
   CommandMode,
   CreateSessionRequest,
-  CreateSessionResponse,
-  DeleteSessionRequest,
-  DeleteSessionResponse,
   ExecuteRequest,
   ExecuteResponse,
   ExecuteStop,
   GetSessionRequest,
-  GetSessionResponse,
-  ListSessionsRequest,
-  ListSessionsResponse,
-  ResolveVarsRequest,
-  ResolveVarsResponse,
   Winsize,
 } from '../grpc/runnerTypes'
-import { IRunnerServiceClient, RunnerServiceClient, RpcError } from '../grpc/client'
+import { IRunnerServiceClient, RpcError } from '../grpc/client'
 import { getSystemShellPath } from '../executors/utils'
 import { IServer } from '../server/runmeServer'
 import { convertEnvList } from '../utils'
 
 import { IRunnerChild, TerminalWindowState } from './types'
 import { GrpcRunnerEnvironment, IRunnerEnvironment } from './environment'
+import { IRunnerClient, GrpcRunnerClient } from './client'
 
 type ExecuteDuplex = DuplexStreamingCall<ExecuteRequest, ExecuteResponse>
 
@@ -169,102 +160,7 @@ export interface IRunnerProgramSession extends IRunnerChild, Pseudoterminal {
   ): void | Promise<void>
 }
 
-type IRunnerClient = IRunnerServiceClient & Disposable
-
-export class GrpcRunnerClient implements IRunnerClient {
-  private disposables: Disposable[] = []
-
-  protected client?: RunnerServiceClient
-  protected _onReady?: EventEmitter<void>
-
-  constructor(
-    protected server: IServer,
-    ready?: EventEmitter<void>,
-  ) {
-    this._onReady = ready
-    this.disposables.push(
-      server.onTransportReady(({ transport }) => this.initRunnerClient(transport)),
-    )
-
-    this.disposables.push(
-      server.onClose(() => {
-        return this.deinitRunnerClient()
-      }),
-    )
-  }
-
-  private deinitRunnerClient() {
-    this.client = undefined
-  }
-
-  private async initRunnerClient(transport?: GrpcTransport) {
-    this.deinitRunnerClient()
-    this.client = new RunnerServiceClient(transport ?? (await this.server.transport()))
-    this._onReady?.fire()
-  }
-
-  static assertClient(client: RunnerServiceClient | undefined): asserts client {
-    if (!client) {
-      throw new Error('Client is not active!')
-    }
-  }
-
-  protected register<T extends Disposable>(d: T): T {
-    this.disposables.push(d)
-    return d
-  }
-
-  async dispose(): Promise<void> {
-    this.disposables.forEach((d) => d.dispose())
-  }
-
-  createSession(
-    input: CreateSessionRequest,
-    options?: RpcOptions | undefined,
-  ): UnaryCall<CreateSessionRequest, CreateSessionResponse> {
-    GrpcRunnerClient.assertClient(this.client)
-    return this.client.createSession(input, options)
-  }
-
-  getSession(
-    input: GetSessionRequest,
-    options?: RpcOptions | undefined,
-  ): UnaryCall<GetSessionRequest, GetSessionResponse> {
-    GrpcRunnerClient.assertClient(this.client)
-    return this.client.getSession(input, options)
-  }
-
-  listSessions(
-    input: ListSessionsRequest,
-    options?: RpcOptions | undefined,
-  ): UnaryCall<ListSessionsRequest, ListSessionsResponse> {
-    GrpcRunnerClient.assertClient(this.client)
-    return this.client.listSessions(input, options)
-  }
-
-  deleteSession(
-    input: DeleteSessionRequest,
-    options?: RpcOptions | undefined,
-  ): UnaryCall<DeleteSessionRequest, DeleteSessionResponse> {
-    GrpcRunnerClient.assertClient(this.client)
-    return this.client.deleteSession(input, options)
-  }
-
-  execute(options?: RpcOptions | undefined): DuplexStreamingCall<ExecuteRequest, ExecuteResponse> {
-    GrpcRunnerClient.assertClient(this.client)
-    return this.client.execute(options)
-  }
-
-  resolveVars(
-    input: ResolveVarsRequest,
-    options?: RpcOptions | undefined,
-  ): UnaryCall<ResolveVarsRequest, ResolveVarsResponse> {
-    GrpcRunnerClient.assertClient(this.client)
-    return this.client.resolveVars(input, options)
-  }
-}
-
-export class GrpcRunner implements IRunner {
+export default class GrpcRunner implements IRunner {
   protected client: IRunnerClient
 
   private children: WeakRef<IRunnerChild>[] = []
